Fix codestyle + package + move data file
