+++ src/data/data.csv
@@ -1,0 +1,4 @@
+event,1,T
+event,get milk,T
+event,get a mobile phone,F
+deadline,get laptop by Mar 3 2020 08:30PM,F
