--- src/main/data/data.csv
@@ -1,4 +1,0 @@
-event,1,T
-deadline,get pen by Feb 2 2020 08:20PM,T
-event,get milk,F
-deadline,get pen by Feb 2 2020 08:20PM,T
