+++ src/main/java/duke/Ui.java
@@ -1,3 +1,4 @@
+package duke;
 public class Ui {
     public static void greet() {
         System.out.println("Greeting, traveler. My name is Andrew. What can I do for you?");
@@ -24,6 +25,17 @@ public class Ui {
 
     public static void printError(Exception e) {
         System.out.println(e.getMessage());
+    }
+
+
+    public static void printTaskList() {
+        System.out.println("Commands available:");
+        System.out.println("1. event [description] : add a new event");
+        System.out.println("2. deadline [description] by yyyy-mm-dd hhmm : add a new deadline");
+        System.out.println("3. todo [description] at yyyy-mm-dd hhmm : add a new todo");
+        System.out.println("4. done [index] : complete the task at given index");
+        System.out.println("5. delete [index] : delete the task at given index");
+        System.out.println("6. bye : say goodbye and quit");
     }
